RandomizedSparsification

Documentation for RandomizedSparsification.

RandomizedSparsification._sparsify_context_data

RandomizedSparsification.r_ε

RandomizedSparsification.rankrand

RandomizedSparsification.representation_error

RandomizedSparsification.sparsify

RandomizedSparsification.srank

RandomizedSparsification.ρ_s

RandomizedSparsification._sparsify_context_data — Method

_sparsify_context_data(A)

Compute the necessary context data, A_1_norm, A_F_norm, P_ij and P_ij_cs (cumulative distribution on the matrix elements) for the sparsify randomized sparisification algorithm of A.

RandomizedSparsification.r_ε — Method

r_ε(A, ε, [rank])
r_ε(d₁, d₂, ε, rank)

Compute the sufficient number of samples r required to achieve a given error tolerance ε in the sparsification of the matrix A ($A \in \mathbb{R}^{d_1 \times d_2}$) in the expected value. Optionally use rank as an upperbound for $\operatorname{srank}$. The number of samples returned by this function is not necessarily the least bound for the number of samples needed for the error to bound to hold.

RandomizedSparsification.rankrand — Method

rankrand(d₁, d₂, rank)

Generate a random matrix of dimensions d₁ × d₂ with a rank rank.

RandomizedSparsification.representation_error — Method

representation_error(A, A_r)

Compute the relative operator representation error between the original matrix A and the sparsified matrix A_r.

\[ \frac{\|A - A_r\|_\text{op}}{\|A\|_\text{op}}\]

RandomizedSparsification.sparsify — Method

sparsify(A; r, multithreaded=false, progress=false)

Sparsify the input matrix A using the randomized element-wise sparsification algorithm with r samples.

RandomizedSparsification.srank — Method

srank(A::AbstractMatrix)

Compute the stable rank of the input matrix A.

The stable rank is defined as the ratio of the squared Frobenius norm to the squared spectral norm of the matrix.

\[ \frac{\|A\|^2_\text{F}}{\|A\|^2_\text{op}}\]

RandomizedSparsification.ρ_s — Method

ρ_s(A::AbstractSparseArray)

Compute the proportional sparsity of the input matrix A.

\[ \frac{\#\{(i,j)\mid A_{ij} \neq 0 \}}{\#\{(i,j)\}}\]