RandomizedSparsification

Implementation of the randomized sparsification algorithm from [1].

Index

RandomizedSparsification.FindIndices

RandomizedSparsification._sparsify_context_data

RandomizedSparsification.r_ε

RandomizedSparsification.rankrand

RandomizedSparsification.representation_error

RandomizedSparsification.sparsify

RandomizedSparsification.srank

RandomizedSparsification.ρ_s

Algorithm

RandomizedSparsification API

RandomizedSparsification.sparsify — Function

sparsify(A; r, multithreaded=false, progress=false)

Sparsify the input matrix A using the randomized element-wise sparsification algorithm with r samples.

RandomizedSparsification.r_ε — Function

r_ε(A, ε, [rank])
r_ε(d₁, d₂, ε, rank)

Compute the sufficient number of samples r required to achieve a given error tolerance ε in the sparsification of the matrix A ($A \in \mathbb{R}^{d_1 \times d_2}$) in the expected value. Optionally use rank as an upperbound for $\operatorname{srank}$. The number of samples returned by this function is not necessarily the least bound for the number of samples needed for the error to bound to hold.

Utility functions

RandomizedSparsification.rankrand — Function

rankrand(d₁, d₂, rank)

Generate a random matrix of dimensions d₁ × d₂ with a rank rank.

RandomizedSparsification.representation_error — Function

representation_error(A, A_r)

Compute the relative operator representation error between the original matrix A and the sparsified matrix A_r.

\[ \frac{\|A - A_r\|_\text{op}}{\|A\|_\text{op}}\]

RandomizedSparsification.srank — Function

srank(A::AbstractMatrix)

Compute the stable rank of the input matrix A.

The stable rank is defined as the ratio of the squared Frobenius norm to the squared spectral norm of the matrix.

\[ \frac{\|A\|^2_\text{F}}{\|A\|^2_\text{op}}\]

RandomizedSparsification.ρ_s — Function

ρ_s(A::AbstractSparseArray)

Compute the proportional sparsity of the input matrix A.

\[ \frac{\#\{(i,j)\mid A_{ij} \neq 0 \}}{\#\{(i,j)\}}\]

Internals

RandomizedSparsification.FindIndices — Module

This module contains functions for determining the indices from the cumulative sum of the probability matrix and the randomly sampled numbers from $[0,1]$ of the length of the desired number of samples.

It does not consider whether or not the number of the samples that are taken fit into the memory or not. This should be considered elsewhere in the call stack where the number of samples desired can be separated into batches that should be sampled at once and only the matrix should be held between the iterations.

The functions that find the summation indices in the sparsification algorithm have the signature signature:

_indices(
    P_ij_cs::AbstractVector{<:Real}, 
    rand01_vec::AbstractVector{<:Real},
    m::Integer, n::Integer
)::AbstractDict{<:CartesianIndex, <:Unsigned}

where P_ij_cs is the vector of the cumulative sum of the probabilities, rand01_vec are the sampled values and m and n are the dimensions of the original matrix. So the indices are returned in the form of a <:AbstractDict{<:CartesianIndex, <:Unsigned}, where the Unsigned holds the counts of the samples of ind::CartesianIndex.

RandomizedSparsification._sparsify_context_data — Function

_sparsify_context_data(A)

Compute the necessary context data, A_1_norm, A_F_norm, P_ij and P_ij_cs (cumulative distribution on the matrix elements) for the sparsify randomized sparisification algorithm of A.

Bibliography

[1]

A. Kundu and P. Drineas. A Note on Randomized Element-wise Matrix Sparsification (2014). Accessed on Jan 6, 2025.